pyproject.toml
[project]
name = "kedro-airflow-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kedro>=1.0.0",
    "kedro-datasets[polars]>=8.1.0",
    "kedro-datasets[aws]>=8.1.0",
    "polars>=1.34.0",
    "kedro-airflow>=0.9.0",
    "kedro-docker>=0.6.0",
    "apache-airflow[amazon]>=2.8.0",
    "boto3>=1.34.0",
    "s3fs>=2024.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.kedro]
package_name = "kedro_airflow_test"
project_name = "kedro_airflow_test"
kedro_init_version = "1.0.0"
source_dir = "src"

[tool.kedro-airflow]
# Airflow configuration will be added here

[tool.kedro-docker]
# Docker image configuration for ECS deployment
image_name = "kedro-airflow-test"
base_image = "python:3.13-slim"
target_image = "kedro-airflow-test:latest"
copy_from = ["src", "conf", "pyproject.toml"]
pip_install_options = ["--no-cache-dir"]
requirements_path = "requirements.txt"

[tool.kedro-docker.env]
KEDRO_ENV = "production"
AWS_REGION = "us-east-1"

[tool.kedro_telemetry]
project_id = "6f4b97f0685b4cbd88f2ba74dcb38855"
